+++ src/main/kotlin/org/hyacinthbots/saffronstatus/_Constants.kt
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package org.hyacinthbots.watchdog
+package org.hyacinthbots.saffronstatus
 
 import com.kotlindiscord.kord.extensions.utils.env
 import com.kotlindiscord.kord.extensions.utils.envOrNull
