+++ src/main/kotlin/org/hyacinthbots/saffronstatus/_Utils.kt
@@ -4,13 +4,13 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package org.hyacinthbots.watchdog
+package org.hyacinthbots.saffronstatus
 
 import com.kotlindiscord.kord.extensions.builders.ExtensibleBotBuilder
 import com.kotlindiscord.kord.extensions.utils.loadModule
 import kotlinx.coroutines.runBlocking
-import org.hyacinthbots.watchdog.database.Database
-import org.hyacinthbots.watchdog.database.collections.WatchedBotCollection
+import org.hyacinthbots.saffronstatus.database.Database
+import org.hyacinthbots.saffronstatus.database.collections.WatchedBotCollection
 import org.koin.dsl.bind
 
 suspend inline fun ExtensibleBotBuilder.database(migrate: Boolean) {
